FIX : Tweaked windows menu to be "standard" ADD : Splash screen FIX : warnings
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@13732 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ GPS/qgps.cpp
@@ -191,7 +191,7 @@ void QGPS::updateGpsStatus()
 	}
 	StrengthView->setSatellites(List);
 	satTracker->setSatellites(List);
-	satTracker->setHeading(gpsDevice->heading());
+    satTracker->setHeading((int)gpsDevice->heading());
 }
 
 void QGPS::showEvent ( QShowEvent * anEvent )
